+++ plugin/src/BeautifulMentionsPlugin.tsx
@@ -8,7 +8,7 @@ import {
   $createTextNode,
   $getSelection,
   $isNodeSelection,
-  $nodesOfType,
+  $isParagraphNode,
   $setSelection,
   BLUR_COMMAND,
   BaseSelection,
@@ -47,6 +47,7 @@ import {
   RENAME_MENTIONS_COMMAND,
 } from "./mention-commands";
 import {
+  $findBeautifulMentionNodes,
   $getSelectionInfo,
   $selectEnd,
   DEFAULT_PUNCTUATION,
@@ -147,6 +148,7 @@ export function BeautifulMentionsPlugin(props: BeautifulMentionsPluginProps) {
   const {
     items,
     onSearch,
+    autoSpace = true,
     searchDelay = props.onSearch ? 250 : 0,
     allowSpaces = true,
     insertOnBlur = true,
@@ -188,7 +190,7 @@ export function BeautifulMentionsPlugin(props: BeautifulMentionsPluginProps) {
     if (!trigger) {
       return [];
     }
-    // Add options from the lookup service
+    // add options from the lookup service
     let opt = results.map((result) => {
       if (typeof result === "string") {
         return new MentionOption(trigger, result, result);
@@ -202,11 +204,11 @@ export function BeautifulMentionsPlugin(props: BeautifulMentionsPluginProps) {
       opt = opt.slice(0, menuItemLimit);
     }
     // Add mentions from the editor. When a search function is provided, wait for the
-    // delayed search to prevent flickering.
+    // delayed search to prevent flickering
     const readyToAddCurrentMentions = !onSearch || (!loading && query !== null);
     if (readyToAddCurrentMentions && showCurrentMentionsAsSuggestions) {
       editor.getEditorState().read(() => {
-        const mentions = $nodesOfType(BeautifulMentionNode);
+        const mentions = $findBeautifulMentionNodes(editor);
         for (const mention of mentions) {
           const value = mention.getValue();
           const data = mention.getData();
@@ -221,7 +223,7 @@ export function BeautifulMentionsPlugin(props: BeautifulMentionsPluginProps) {
         }
       });
     }
-    // Add option to create a new mention
+    // add option to create a new mention
     if (query && opt.every((o) => o.displayValue !== query)) {
       const displayValue =
         typeof creatable === "string"
@@ -275,7 +277,14 @@ export function BeautifulMentionsPlugin(props: BeautifulMentionsPluginProps) {
           selectedOption.data,
         );
         if (nodeToReplace) {
+          const nextSibling = nodeToReplace.getNextSibling();
           nodeToReplace.replace(mentionNode);
+          if (nextSibling instanceof TextNode) {
+            const nextSiblingTextContent = nextSibling.getTextContent();
+            if (!/\s/.test(nextSiblingTextContent)) {
+              nextSibling.remove();
+            }
+          }
         }
         closeMenu?.();
         justSelectedAnOption.current = true;
@@ -306,12 +315,6 @@ export function BeautifulMentionsPlugin(props: BeautifulMentionsPluginProps) {
 
   const checkForMentionMatch = useCallback(
     (text: string) => {
-      // Don't show the menu if the next character is a word character
-      const selectionInfo = $getSelectionInfo(triggers, punctuation);
-      if (selectionInfo?.isTextNode && selectionInfo.wordCharAfterCursor) {
-        return null;
-      }
-
       const queryMatch = checkForMentions(
         text,
         triggers,
@@ -365,7 +368,7 @@ export function BeautifulMentionsPlugin(props: BeautifulMentionsPluginProps) {
       triggers,
       preTriggerChars,
       punctuation,
-      false,
+      allowSpaces,
     );
     if (queryMatch === null) {
       return false;
@@ -390,7 +393,15 @@ export function BeautifulMentionsPlugin(props: BeautifulMentionsPluginProps) {
       { tag: "history-merge" },
     );
     return true;
-  }, [editor, options, preTriggerChars, punctuation, trigger, triggers]);
+  }, [
+    editor,
+    options,
+    preTriggerChars,
+    punctuation,
+    trigger,
+    triggers,
+    allowSpaces,
+  ]);
 
   const restoreSelection = useCallback(() => {
     const selection = $getSelection();
@@ -432,13 +443,20 @@ export function BeautifulMentionsPlugin(props: BeautifulMentionsPluginProps) {
 
   const insertSpaceIfNecessary = useCallback(
     (startsWithTriggerChar = false) => {
+      if (!autoSpace) {
+        return;
+      }
+
       const selectionInfo = $getSelectionInfo(triggers, punctuation);
       if (!selectionInfo) {
-        return false;
+        return;
       }
+
       const {
         node,
         offset,
+        type,
+        parentNode,
         isTextNode,
         textContent,
         prevNode,
@@ -447,51 +465,82 @@ export function BeautifulMentionsPlugin(props: BeautifulMentionsPluginProps) {
         cursorAtStartOfNode,
         cursorAtEndOfNode,
       } = selectionInfo;
-      // [Mention][|][Text]
+
+      // [Mention][|][ Text]
       if (
         isTextNode &&
         cursorAtStartOfNode &&
         $isBeautifulMentionNode(prevNode)
       ) {
         node.insertBefore($createTextNode(" "));
+        return;
       }
-      // [Text][|][Mention]
+
+      // ^[|][Mention]
+      if (
+        $isBeautifulMentionNode(node) &&
+        prevNode === null &&
+        $isParagraphNode(parentNode) &&
+        type === "element" &&
+        offset === 0
+      ) {
+        const textNode = $createTextNode(" ");
+        node.insertBefore(textNode);
+        textNode.selectStart();
+        return;
+      }
+
+      // [Text ][|][Mention]
       if (
         isTextNode &&
         cursorAtEndOfNode &&
         $isBeautifulMentionNode(nextNode)
       ) {
         node.insertAfter($createTextNode(" "));
+        return;
       }
-      // [Text][|][Word]
+
+      // [Text ][|][Text]
       if (isTextNode && startsWithTriggerChar && wordCharAfterCursor) {
         const content =
           textContent.substring(0, offset) +
           " " +
           textContent.substring(offset);
         node.setTextContent(content);
+        return;
       }
+
       // [Mention][|]
       if ($isBeautifulMentionNode(node) && nextNode === null) {
         node.insertAfter($createTextNode(" "));
       }
     },
-    [punctuation, triggers],
+    [punctuation, triggers, autoSpace],
   );
 
   const handleKeyDown = useCallback(
     (event: KeyboardEvent) => {
       const { key, metaKey, ctrlKey } = event;
-      const simpleKey = key.length === 1;
+      const simpleKey = key?.length === 1;
       const isTrigger = triggers.some((trigger) => key === trigger);
       const wordChar = isWordChar(key, triggers, punctuation);
-      if (!simpleKey || (!wordChar && !isTrigger) || metaKey || ctrlKey) {
+      const isSpace = allowSpaces && /^\s$/.test(key);
+      if (!simpleKey || metaKey || ctrlKey) {
         return false;
+      }
+      if (!wordChar && !isTrigger && !isSpace) {
+        return convertTextToMention();
       }
       insertSpaceIfNecessary(isTrigger);
       return false;
     },
-    [insertSpaceIfNecessary, punctuation, triggers],
+    [
+      insertSpaceIfNecessary,
+      punctuation,
+      convertTextToMention,
+      triggers,
+      allowSpaces,
+    ],
   );
 
   const handlePaste = useCallback(
@@ -502,6 +551,7 @@ export function BeautifulMentionsPlugin(props: BeautifulMentionsPluginProps) {
       const isPunctuation =
         firstChar && new RegExp(`[\\s${punctuation}]`).test(firstChar);
       if (isTrigger || !isPunctuation) {
+        // insert space before pasting if the content starts with a trigger character
         insertSpaceIfNecessary();
       }
       return false; // will be handled by the lexical clipboard module
@@ -522,6 +572,8 @@ export function BeautifulMentionsPlugin(props: BeautifulMentionsPluginProps) {
           const selection = $getSelection();
           if (selection && !$isNodeSelection(selection)) {
             oldSelection.current = selection;
+          } else if (!selection) {
+            oldSelection.current = null;
           }
           return false;
         },
@@ -578,7 +630,15 @@ export function BeautifulMentionsPlugin(props: BeautifulMentionsPluginProps) {
       ),
       editor.registerCommand(
         REMOVE_MENTIONS_COMMAND,
-        ({ trigger, value, focus }) => $removeMention(trigger, value, focus),
+        ({ trigger, value, focus }) => {
+          const removed = $removeMention(trigger, value, focus);
+          if (!focus) {
+            // remove oldSelection manually because the SELECTION_CHANGE_COMMAND
+            // listener is not called in this case.
+            oldSelection.current = null;
+          }
+          return removed;
+        },
         COMMAND_PRIORITY_LOW,
       ),
       editor.registerCommand(