+++ plugin/src/MentionComponent.tsx
@@ -240,6 +240,7 @@ export default function BeautifulMentionComponent(
   if (Component) {
     return (
       <Component
+        // @ts-ignore
         ref={ref}
         trigger={trigger}
         value={value}